Fix device not working after restart as initial report not sent
The device now sends always sends all reports on usb ready. This was being
blocked by the logic to prevent repeat reports being sent once established.

This replaces the detection method for sending each report to a bit mask
of buttons that changed state, which makes it much easier to work with.

The system control reports have been removed as these are largely not
supported by Linux or Windows, so aren't much use.